labs/capstone-options/option-a-code-review/python/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
anthropic>=0.40.0
python-dotenv>=1.0.0
slowapi>=0.1.9
